adapt test_dataset to new structure

+++ tests/test_dataset.py
@@ -4,38 +4,39 @@ from gooddataclient.project import Project, delete_projects_by_name
 from gooddataclient.connection import Connection
 from gooddataclient.dataset import DateDimension
 
-from tests.credentials import password, username
+from tests.credentials import password, username, project_id
 from tests.test_project import TEST_PROJECT_NAME
 from tests import logger, examples
+
 
 class TestDataset(unittest.TestCase):
 
     def setUp(self):
-        self.connection = Connection(username, password, debug=0)
-        #drop all the test projects:
-        delete_projects_by_name(self.connection, TEST_PROJECT_NAME)
-        self.project = Project(self.connection).create(TEST_PROJECT_NAME)
+        self.connection = Connection(username, password)
+
+        if not project_id:
+            delete_projects_by_name(self.connection, TEST_PROJECT_NAME)
+            self.project = Project(self.connection).create(TEST_PROJECT_NAME)
+        else:
+            self.project = Project(self.connection).load(id=project_id)
 
     def tearDown(self):
-        self.project.delete()
+        if not project_id:
+            self.project.delete()
 
-    def test_create_date_dimension(self):
+    def test_create_dataset(self):
         for (example, ExampleDataset) in examples.examples:
             dataset = ExampleDataset(self.project)
-            date_dimension = dataset.get_date_dimension()
-            if date_dimension:
-                DateDimension(self.project).create(name=date_dimension.schemaReference,
-                               include_time=date_dimension.datetime)
-                # TODO: verify the creation
+            dataset.create()
+            # TODO: verify the creation
 
     def test_upload_dataset(self):
         for (example, ExampleDataset) in examples.examples:
             dataset = ExampleDataset(self.project)
             dataset.upload()
             dataset_metadata = dataset.get_metadata(name=dataset.schema_name)
-            self.assert_(dataset_metadata['dataUploads'])
+            self.assertTrue(dataset_metadata['dataUploads'])
             self.assertEquals('OK', dataset_metadata['lastUpload']['dataUploadShort']['status'])
-            dataset.upload()
             # TODO: check different data for the upload
 
     def test_date_maql(self):
